added workout validation and delete category validation

+++ workout/src/main/java/com/ibm/workout/Service/CategoryService.java
@@ -1,17 +1,22 @@
 package com.ibm.workout.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.ibm.workout.entity.Category;
+import com.ibm.workout.entity.Workout;
 import com.ibm.workout.repository.CategoryRepository;
+import com.ibm.workout.repository.WorkoutRepository;
 
 @Service
 public class CategoryService {
 	@Autowired
 	CategoryRepository categoryRepository;
+	@Autowired
+	WorkoutRepository workoutRepository;
 
 	public String createCategory(Category category) {
 		String title = category.getTitle();
@@ -39,7 +44,17 @@ public class CategoryService {
 	}
 
 	public void deleteCategory(String categoryId) {
-		categoryRepository.deleteById(categoryId);
+		Optional<Category> result = categoryRepository.findById(categoryId);
+		if(result.isPresent()) {
+			String title = result.get().getTitle();
+			List<Workout> workoutResult = workoutRepository.findByCategoryIgnoreCase(title);
+			if (workoutResult.isEmpty()) {
+				categoryRepository.deleteById(categoryId);
+			} else {
+				throw new CustomIllegalArgumentException("This category is used by a workout");
+			}
+		}
+		
 	}
 
 }
